Refactored Periodic into periodic.py; refactored Stationary to bring back K_r2 vs K_r functionality
