+++ .gitignore
@@ -67,3 +67,6 @@ target/
 
 # VSCode
 .vscode
+
+# OSX
+.DS_Store
